% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CANCERGOV_DRUG}
\alias{CANCERGOV_DRUG}
\title{CANCERGOV_DRUG}
\format{
A data frame with 67835 rows and 5 variables:
\describe{
\item{\code{cd_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{drug_link}}{character COLUMN_DESCRIPTION}
\item{\code{ncit_code}}{character COLUMN_DESCRIPTION}
\item{\code{drug_synonym_type}}{character COLUMN_DESCRIPTION}
\item{\code{cancergov_drug}}{character COLUMN_DESCRIPTION}
}
}
\usage{
CANCERGOV_DRUG
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
