% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONCEPT_DEFINITION}
\alias{CONCEPT_DEFINITION}
\title{CONCEPT_DEFINITION}
\format{
A data frame with 7804 rows and 3 variables:
\describe{
\item{\code{concept_id}}{double COLUMN_DESCRIPTION}
\item{\code{concept_name}}{character COLUMN_DESCRIPTION}
\item{\code{concept_definition}}{character COLUMN_DESCRIPTION}
}
}
\usage{
CONCEPT_DEFINITION
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
