% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRUG_DICTIONARY}
\alias{DRUG_DICTIONARY}
\title{DRUG_DICTIONARY}
\format{
A data frame with 8156 rows and 3 variables:
\describe{
\item{\code{dd_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{drug}}{character COLUMN_DESCRIPTION}
\item{\code{definition}}{character COLUMN_DESCRIPTION}
}
}
\usage{
DRUG_DICTIONARY
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
