% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRUG_DICTIONARY_LOG}
\alias{DRUG_DICTIONARY_LOG}
\title{DRUG_DICTIONARY_LOG}
\format{
A data frame with 12 rows and 2 variables:
\describe{
\item{\code{ddl_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{drug_count}}{double COLUMN_DESCRIPTION}
}
}
\usage{
DRUG_DICTIONARY_LOG
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
