% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRUG_LINK}
\alias{DRUG_LINK}
\title{DRUG_LINK}
\format{
A data frame with 7785 rows and 3 variables:
\describe{
\item{\code{dl_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{drug}}{character COLUMN_DESCRIPTION}
\item{\code{drug_link}}{character COLUMN_DESCRIPTION}
}
}
\usage{
DRUG_LINK
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
