% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRUG_LINK_NCIT}
\alias{DRUG_LINK_NCIT}
\title{DRUG_LINK_NCIT}
\format{
A data frame with 54708 rows and 3 variables:
\describe{
\item{\code{dln_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{drug_link}}{character COLUMN_DESCRIPTION}
\item{\code{ncit_code}}{character COLUMN_DESCRIPTION}
}
}
\usage{
DRUG_LINK_NCIT
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
