% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRUG_LINK_SYNONYM}
\alias{DRUG_LINK_SYNONYM}
\title{DRUG_LINK_SYNONYM}
\format{
A data frame with 18944 rows and 4 variables:
\describe{
\item{\code{dls_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{drug_link}}{character COLUMN_DESCRIPTION}
\item{\code{drug_synonym_type}}{character COLUMN_DESCRIPTION}
\item{\code{drug_synonym}}{character COLUMN_DESCRIPTION}
}
}
\usage{
DRUG_LINK_SYNONYM
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
