% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRUG_LINK_URL}
\alias{DRUG_LINK_URL}
\title{DRUG_LINK_URL}
\format{
A data frame with 23933 rows and 3 variables:
\describe{
\item{\code{dlu_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{drug_link}}{character COLUMN_DESCRIPTION}
\item{\code{drug_link_url}}{character COLUMN_DESCRIPTION}
}
}
\usage{
DRUG_LINK_URL
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
