% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCIT_SYNONYM}
\alias{NCIT_SYNONYM}
\title{NCIT_SYNONYM}
\format{
A data frame with 49791 rows and 7 variables:
\describe{
\item{\code{ns_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{ncit_code}}{character COLUMN_DESCRIPTION}
\item{\code{ncit_code_url}}{character COLUMN_DESCRIPTION}
\item{\code{term}}{character COLUMN_DESCRIPTION}
\item{\code{source}}{character COLUMN_DESCRIPTION}
\item{\code{tty}}{character COLUMN_DESCRIPTION}
\item{\code{code}}{character COLUMN_DESCRIPTION}
}
}
\usage{
NCIT_SYNONYM
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
