% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEW_CONCEPTS}
\alias{NEW_CONCEPTS}
\title{NEW_CONCEPTS}
\format{
A data frame with 50 rows and 3 variables:
\describe{
\item{\code{component}}{character COLUMN_DESCRIPTION}
\item{\code{concept_id}}{character COLUMN_DESCRIPTION}
\item{\code{new_concept_name}}{character COLUMN_DESCRIPTION}
}
}
\usage{
NEW_CONCEPTS
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
