% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TEMP_MAPPING}
\alias{TEMP_MAPPING}
\title{TEMP_MAPPING}
\format{
A data frame with 49 rows and 7 variables:
\describe{
\item{\code{rowid}}{character COLUMN_DESCRIPTION}
\item{\code{component}}{character COLUMN_DESCRIPTION}
\item{\code{new_concept_name}}{character COLUMN_DESCRIPTION}
\item{\code{new_concept_words}}{character COLUMN_DESCRIPTION}
\item{\code{drug_link}}{character COLUMN_DESCRIPTION}
\item{\code{drug_link_name_match}}{character COLUMN_DESCRIPTION}
\item{\code{drug_link_word_match}}{character COLUMN_DESCRIPTION}
}
}
\usage{
TEMP_MAPPING
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
