% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{V20200928230401}
\alias{V20200928230401}
\title{V20200928230401}
\format{
A data frame with 980 rows and 12 variables:
\describe{
\item{\code{routine_id}}{character COLUMN_DESCRIPTION}
\item{\code{missing_drug}}{character COLUMN_DESCRIPTION}
\item{\code{apr_generic_name}}{character COLUMN_DESCRIPTION}
\item{\code{apr_investigative_flag}}{character COLUMN_DESCRIPTION}
\item{\code{cdg_common_generic_name}}{character COLUMN_DESCRIPTION}
\item{\code{notes}}{character COLUMN_DESCRIPTION}
\item{\code{fact}}{character COLUMN_DESCRIPTION}
\item{\code{apr_generic_name_longest_word_exact}}{character COLUMN_DESCRIPTION}
\item{\code{apr_generic_name_longest_word_left_join_status}}{character COLUMN_DESCRIPTION}
\item{\code{apr_generic_name_longest_word_left_join}}{character COLUMN_DESCRIPTION}
\item{\code{processing_type}}{character COLUMN_DESCRIPTION}
\item{\code{concept}}{character COLUMN_DESCRIPTION}
}
}
\usage{
V20200928230401
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
